+++ yandex/data_source_yandex_mdb_mysql_cluster.go
@@ -257,6 +257,10 @@ func dataSourceYandexMDBMySQLCluster() *schema.Resource {
 							Type:     schema.TypeBool,
 							Computed: true,
 						},
+						"data_transfer": {
+							Type:     schema.TypeBool,
+							Computed: true,
+						},
 					},
 				},
 			},