+++ yandex/data_source_yandex_mdb_postgresql_cluster.go
@@ -114,6 +114,10 @@ func dataSourceYandexMDBPostgreSQLClusterConfigBlock() *schema.Resource {
 							Type:     schema.TypeBool,
 							Computed: true,
 						},
+						"data_transfer": {
+							Type:     schema.TypeBool,
+							Computed: true,
+						},
 					},
 				},
 			},