+++ yandex/mdb_mysql_structures_test.go
@@ -4,11 +4,10 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/golang/protobuf/ptypes/wrappers"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/golang/protobuf/ptypes/wrappers"
 	"github.com/yandex-cloud/go-genproto/yandex/cloud/mdb/mysql/v1"
-
 	config "github.com/yandex-cloud/go-genproto/yandex/cloud/mdb/mysql/v1/config"
 )